ci(lint): fix golangci-lint v1.64 findings
- Add doc comments to exported Conn/Stream methods (revive).
- Rename unused params to _ in test fakes and TLS test handler.
- Drop unnecessary type conversions in frame and hpack
  (unconvert).
- Drop blank-identifier read in conn_test.go (gosimple S1005).

No behavior change. Phase A unconvert fixes are the same casts
that v1.62 silently accepted; v1.64 surfaces them.

+++ hpack/dynamic_table.go
@@ -89,8 +89,8 @@ func (d *dynamicTable) evictOldest() {
 		return
 	}
 	e := d.entries[d.head]
-	d.size -= uint32(e.nameLen) + uint32(e.valueLen) + 32
-	d.used -= uint32(e.nameLen) + uint32(e.valueLen)
+	d.size -= e.nameLen + e.valueLen + 32
+	d.used -= e.nameLen + e.valueLen
 	d.head = (d.head + 1) % len(d.entries)
 	d.count--
 	if d.count == 0 {
